refactor(models): models/git 을 도메인 모듈 10종으로 분할
991줄 단일 파일의 최상위 정의 81종을 10개 모듈로 옮겼다 — branches 154 ·
repository 189 · working_tree 101 · merge 99 · github 82 · commits 76 ·
remotes 69 · merge_requests 62 · enums 57 · permissions 57.
__init__.py(191)는 재노출만 하며 `from models.git import X` 는 분할 전과
동일하게 동작한다. 도메인 구획은 api/git/ 패키지(B1)와 대칭이다.

**GIT_REPOSITORIES 와 그것을 읽고 쓰는 함수 6종을 같은 모듈에 뒀다.**
이 딕셔너리는 모듈 레벨 가변 상태이고 함수들이 global 로 재바인딩하므로,
다른 모듈로 가르면 사본이 분열돼 한쪽이 등록한 저장소를 다른 쪽이 못 본다.
ruff·mypy 는 그 상태를 통과시키고 테스트도 한쪽 경로만 타면 통과한다.
계획서가 이것을 B5 의 하중 지지대로 지목한 이유다.

**분할을 sed 라인 슬라이스가 아니라 AST 이름 기반으로 했다.** 도메인별로
묶으면 정의가 파일 곳곳에 흩어져 라인 산술이 위험하고, 무엇보다 추출과
검증이 같은 원리를 쓰면 자기확인이 된다(B2 claude_sessions 에서 실제로
`return results` 유실을 IDENTICAL 로 오판했다). 추출 스크립트에 **커버리지
단언**을 넣어 최상위 이름 하나라도 배정 누락·중복이면 즉시 실패하도록 했다 —
유실을 사후에 찾는 대신 불가능하게 만든다.

모듈별 import 는 정의가 실제로 참조하는 이름에서 역산했다(눈으로 훑지 않음).
B4 에서 눈으로 훑다 formatTrendData 를 놓친 사례가 있다.

게이트:
- split_audit.py (ff20c2f 기준) — 정의 77 + 모듈할당 4, 유실·추가·중복·
  본문불일치 **0건**. ruff format 적용 후 재실행해도 0건
- ruff check / ruff format --check — 통과
- mypy — Success: no issues found in 299 source files
- pytest — `1 failed, 1357 passed, 2 skipped` = **베이스라인과 동일**.
  그 1건은 로컬 .env 의 RAG_EMBEDDING_MODEL 오버라이드에서 오는 알려진
  플레이크이며 CI 는 통과한다
